requirements.txt
psycopg2-binary~=2.9.10
yfinance~=0.2.61
pandas~=2.2.3
tqdm~=4.67.1
dotenv~=0.9.9
python-dotenv~=1.1.0